+++ include/camodocal/camera_models/ScaramuzzaCamera.h
@@ -4,11 +4,10 @@
 #include <opencv2/core/core.hpp>
 #include <string>
 
-#include "ceres/rotation.h"
 #include "Camera.h"
+#include "ceres/rotation.h"
 
-namespace camodocal
-{
+namespace camodocal {
 
 #define SCARAMUZZA_POLY_SIZE 5
 #define SCARAMUZZA_INV_POLY_SIZE 20
@@ -20,41 +19,67 @@ namespace camodocal
  * https://sites.google.com/site/scarabotix/ocamcalib-toolbox
  */
 
-class OCAMCamera: public Camera
-{
-public:
-    class Parameters: public Camera::Parameters
-    {
-    public:
+class OCAMCamera : public Camera {
+  public:
+    class Parameters : public Camera::Parameters {
+      public:
         Parameters();
 
-        double& C(void) { return m_C; }
-        double& D(void) { return m_D; }
-        double& E(void) { return m_E; }
+        double& C(void) {
+            return m_C;
+        }
+        double& D(void) {
+            return m_D;
+        }
+        double& E(void) {
+            return m_E;
+        }
 
-        double& center_x(void) { return m_center_x; }
-        double& center_y(void) { return m_center_y; }
+        double& center_x(void) {
+            return m_center_x;
+        }
+        double& center_y(void) {
+            return m_center_y;
+        }
 
-        double& poly(int idx) { return m_poly[idx]; }
-        double& inv_poly(int idx) { return m_inv_poly[idx]; }
+        double& poly(int idx) {
+            return m_poly[idx];
+        }
+        double& inv_poly(int idx) {
+            return m_inv_poly[idx];
+        }
 
-        double C(void) const { return m_C; }
-        double D(void) const { return m_D; }
-        double E(void) const { return m_E; }
+        double C(void) const {
+            return m_C;
+        }
+        double D(void) const {
+            return m_D;
+        }
+        double E(void) const {
+            return m_E;
+        }
 
-        double center_x(void) const { return m_center_x; }
-        double center_y(void) const { return m_center_y; }
+        double center_x(void) const {
+            return m_center_x;
+        }
+        double center_y(void) const {
+            return m_center_y;
+        }
 
-        double poly(int idx) const { return m_poly[idx]; }
-        double inv_poly(int idx) const { return m_inv_poly[idx]; }
+        double poly(int idx) const {
+            return m_poly[idx];
+        }
+        double inv_poly(int idx) const {
+            return m_inv_poly[idx];
+        }
 
         bool readFromYamlFile(const std::string& filename);
         void writeToYamlFile(const std::string& filename) const;
 
         Parameters& operator=(const Parameters& other);
-        friend std::ostream& operator<< (std::ostream& out, const Parameters& params);
+        friend std::ostream& operator<<(std::ostream& out, const Parameters& params);
 
-    private:
+      private:
         double m_poly[SCARAMUZZA_POLY_SIZE];
         double m_inv_poly[SCARAMUZZA_INV_POLY_SIZE];
         double m_C;
@@ -67,8 +92,8 @@ public:
     OCAMCamera();
 
     /**
-    * \brief Constructor from the projection model parameters
-    */
+     * \brief Constructor from the projection model parameters
+     */
     OCAMCamera(const Parameters& params);
 
     Camera::ModelType modelType(void) const;
@@ -76,9 +101,9 @@ public:
     int imageWidth(void) const;
     int imageHeight(void) const;
 
-    void estimateIntrinsics(const cv::Size& boardSize,
-                            const std::vector< std::vector<cv::Point3f> >& objectPoints,
-                            const std::vector< std::vector<cv::Point2f> >& imagePoints);
+    void estimateIntrinsics(
+        const cv::Size& boardSize, const std::vector<std::vector<cv::Point3f> >& objectPoints,
+        const std::vector<std::vector<cv::Point2f> >& imagePoints);
 
     // Lift points from the image plane to the sphere
     void liftSphere(const Eigen::Vector2d& p, Eigen::Vector3d& P) const;
@@ -94,7 +119,7 @@ public:
 
     // Projects 3D points to the image plane (Pi function)
     // and calculates jacobian
-    //void spaceToPlane(const Eigen::Vector3d& P, Eigen::Vector2d& p,
+    // void spaceToPlane(const Eigen::Vector3d& P, Eigen::Vector2d& p,
     //                  Eigen::Matrix<double,2,3>& J) const;
     //%output p
     //%output J
@@ -103,33 +128,25 @@ public:
     //%output p
 
     template <typename T>
-    static void spaceToPlane(const T* const params,
-                             const T* const q, const T* const t,
-                             const Eigen::Matrix<T, 3, 1>& P,
-                             Eigen::Matrix<T, 2, 1>& p);
+    static void spaceToPlane(
+        const T* const params, const T* const q, const T* const t, const Eigen::Matrix<T, 3, 1>& P,
+        Eigen::Matrix<T, 2, 1>& p);
     template <typename T>
-    static void spaceToSphere(const T* const params,
-                              const T* const q, const T* const t,
-                              const Eigen::Matrix<T, 3, 1>& P,
-                              Eigen::Matrix<T, 3, 1>& P_s);
+    static void spaceToSphere(
+        const T* const params, const T* const q, const T* const t, const Eigen::Matrix<T, 3, 1>& P,
+        Eigen::Matrix<T, 3, 1>& P_s);
     template <typename T>
-    static void LiftToSphere(const T* const params,
-                              const Eigen::Matrix<T, 2, 1>& p,
-                              Eigen::Matrix<T, 3, 1>& P);
+    static void LiftToSphere(const T* const params, const Eigen::Matrix<T, 2, 1>& p, Eigen::Matrix<T, 3, 1>& P);
 
     template <typename T>
-    static void SphereToPlane(const T* const params, const Eigen::Matrix<T, 3, 1>& P,
-                               Eigen::Matrix<T, 2, 1>& p);
-
+    static void SphereToPlane(const T* const params, const Eigen::Matrix<T, 3, 1>& P, Eigen::Matrix<T, 2, 1>& p);
 
     void initUndistortMap(cv::Mat& map1, cv::Mat& map2, double fScale = 1.0) const;
-    cv::Mat initUndistortRectifyMap(cv::Mat& map1, cv::Mat& map2,
-                                    float fx = -1.0f, float fy = -1.0f,
-                                    cv::Size imageSize = cv::Size(0, 0),
-                                    float cx = -1.0f, float cy = -1.0f,
-                                    cv::Mat rmat = cv::Mat::eye(3, 3, CV_32F)) const;
+    cv::Mat initUndistortRectifyMap(
+        cv::Mat& map1, cv::Mat& map2, float fx = -1.0f, float fy = -1.0f, cv::Size imageSize = cv::Size(0, 0),
+        float cx = -1.0f, float cy = -1.0f, cv::Mat rmat = cv::Mat::eye(3, 3, CV_32F)) const;
 
-    int parameterCount(void) const;
+    size_t parameterCount(void) const;
 
     const Parameters& getParameters(void) const;
     void setParameters(const Parameters& parameters);
@@ -141,7 +158,7 @@ public:
 
     std::string parametersToString(void) const;
 
-private:
+  private:
     Parameters mParameters;
 
     double m_inv_scale;
@@ -151,12 +168,9 @@ typedef boost::shared_ptr<OCAMCamera> OCAMCameraPtr;
 typedef boost::shared_ptr<const OCAMCamera> OCAMCameraConstPtr;
 
 template <typename T>
-void
-OCAMCamera::spaceToPlane(const T* const params,
-                         const T* const q, const T* const t,
-                         const Eigen::Matrix<T, 3, 1>& P,
-                         Eigen::Matrix<T, 2, 1>& p)
-{
+void OCAMCamera::spaceToPlane(
+    const T* const params, const T* const q, const T* const t, const Eigen::Matrix<T, 3, 1>& P,
+    Eigen::Matrix<T, 2, 1>& p) {
     T P_c[3];
     {
         T P_w[3];
@@ -178,14 +192,14 @@ OCAMCamera::spaceToPlane(const T* const params,
     T c = params[0];
     T d = params[1];
     T e = params[2];
-    T xc[2] = { params[3], params[4] };
+    T xc[2] = {params[3], params[4]};
 
-    //T poly[SCARAMUZZA_POLY_SIZE];
-    //for (int i=0; i < SCARAMUZZA_POLY_SIZE; i++)
+    // T poly[SCARAMUZZA_POLY_SIZE];
+    // for (int i=0; i < SCARAMUZZA_POLY_SIZE; i++)
     //    poly[i] = params[5+i];
 
     T inv_poly[SCARAMUZZA_INV_POLY_SIZE];
-    for (int i=0; i < SCARAMUZZA_INV_POLY_SIZE; i++)
+    for (int i = 0; i < SCARAMUZZA_INV_POLY_SIZE; i++)
         inv_poly[i] = params[5 + SCARAMUZZA_POLY_SIZE + i];
 
     T norm_sqr = P_c[0] * P_c[0] + P_c[1] * P_c[1];
@@ -197,29 +211,22 @@ OCAMCamera::spaceToPlane(const T* const params,
     T rho = T(0.0);
     T theta_i = T(1.0);
 
-    for (int i = 0; i < SCARAMUZZA_INV_POLY_SIZE; i++)
-    {
+    for (int i = 0; i < SCARAMUZZA_INV_POLY_SIZE; i++) {
         rho += theta_i * inv_poly[i];
         theta_i *= theta;
     }
 
     T invNorm = T(1.0) / norm;
-    T xn[2] = {
-        P_c[0] * invNorm * rho,
-        P_c[1] * invNorm * rho
-    };
+    T xn[2] = {P_c[0] * invNorm * rho, P_c[1] * invNorm * rho};
 
     p(0) = xn[0] * c + xn[1] * d + xc[0];
-    p(1) = xn[0] * e + xn[1]     + xc[1];
+    p(1) = xn[0] * e + xn[1] + xc[1];
 }
 
 template <typename T>
-void
-OCAMCamera::spaceToSphere(const T* const params,
-                          const T* const q, const T* const t,
-                          const Eigen::Matrix<T, 3, 1>& P,
-                          Eigen::Matrix<T, 3, 1>& P_s)
-{
+void OCAMCamera::spaceToSphere(
+    const T* const params, const T* const q, const T* const t, const Eigen::Matrix<T, 3, 1>& P,
+    Eigen::Matrix<T, 3, 1>& P_s) {
     T P_c[3];
     {
         T P_w[3];
@@ -238,8 +245,8 @@ OCAMCamera::spaceToSphere(const T* const params,
         P_c[2] += t[2];
     }
 
-    //T poly[SCARAMUZZA_POLY_SIZE];
-    //for (int i=0; i < SCARAMUZZA_POLY_SIZE; i++)
+    // T poly[SCARAMUZZA_POLY_SIZE];
+    // for (int i=0; i < SCARAMUZZA_POLY_SIZE; i++)
     //    poly[i] = params[5+i];
 
     T norm_sqr = P_c[0] * P_c[0] + P_c[1] * P_c[1] + P_c[2] * P_c[2];
@@ -253,25 +260,21 @@ OCAMCamera::spaceToSphere(const T* const params,
 }
 
 template <typename T>
-void
-OCAMCamera::LiftToSphere(const T* const params,
-                          const Eigen::Matrix<T, 2, 1>& p,
-                          Eigen::Matrix<T, 3, 1>& P)
-{
+void OCAMCamera::LiftToSphere(const T* const params, const Eigen::Matrix<T, 2, 1>& p, Eigen::Matrix<T, 3, 1>& P) {
     T c = params[0];
     T d = params[1];
     T e = params[2];
-    T cc[2] = { params[3], params[4] };
+    T cc[2] = {params[3], params[4]};
     T poly[SCARAMUZZA_POLY_SIZE];
-    for (int i=0; i < SCARAMUZZA_POLY_SIZE; i++)
-       poly[i] = params[5+i];
+    for (int i = 0; i < SCARAMUZZA_POLY_SIZE; i++)
+        poly[i] = params[5 + i];
 
     // Relative to Center
     T p_2d[2];
     p_2d[0] = T(p(0));
     p_2d[1] = T(p(1));
 
-    T xc[2] = { p_2d[0] - cc[0], p_2d[1] - cc[1]};
+    T xc[2] = {p_2d[0] - cc[0], p_2d[1] - cc[1]};
 
     T inv_scale = T(1.0) / (c - d * e);
 
@@ -286,9 +289,8 @@ OCAMCamera::LiftToSphere(const T* const params,
     T phi_i = T(1.0);
     T z = T(0.0);
 
-    for (int i = 0; i < SCARAMUZZA_POLY_SIZE; i++)
-    {
-        if (i!=1) {
+    for (int i = 0; i < SCARAMUZZA_POLY_SIZE; i++) {
+        if (i != 1) {
             z += phi_i * poly[i];
         }
         phi_i *= phi;
@@ -306,8 +308,7 @@ OCAMCamera::LiftToSphere(const T* const params,
 }
 
 template <typename T>
-void OCAMCamera::SphereToPlane(const T* const params, const Eigen::Matrix<T, 3, 1>& P,
-                               Eigen::Matrix<T, 2, 1>& p) {
+void OCAMCamera::SphereToPlane(const T* const params, const Eigen::Matrix<T, 3, 1>& P, Eigen::Matrix<T, 2, 1>& p) {
     T P_c[3];
     {
         P_c[0] = T(P(0));
@@ -326,7 +327,8 @@ void OCAMCamera::SphereToPlane(const T* const params, const Eigen::Matrix<T, 3, 
 
     T norm_sqr = P_c[0] * P_c[0] + P_c[1] * P_c[1];
     T norm = T(0.0);
-    if (norm_sqr > T(0.0)) norm = sqrt(norm_sqr);
+    if (norm_sqr > T(0.0))
+        norm = sqrt(norm_sqr);
 
     T theta = atan2(-P_c[2], norm);
     T rho = T(0.0);
@@ -343,6 +345,6 @@ void OCAMCamera::SphereToPlane(const T* const params, const Eigen::Matrix<T, 3, 
     p(0) = xn[0] * c + xn[1] * d + xc[0];
     p(1) = xn[0] * e + xn[1] + xc[1];
 }
-}
+}  // namespace camodocal
 
 #endif
